Slice up snowflake.go
Move config and node related code into their own files.
